remove autopopulation for getOneProject simplify createTicket delete tasks.js and relocate tickets.js into projects update ticket routes
