--- routes/project/tasks.js
@@ -1,11 +1,0 @@
-import { Router } from 'express';
-// import controllers
-
-const router = Router();
-
-// task routes
-// CREATE new task
-// UPDATE Task status
-// UPDATE Task assignee
-
-export default router
